+++ src/main/java/de/skillkiller/documentdbackend/entity/http/frontend/response/UpdateDocumentResponse.java
@@ -1,4 +1,4 @@
-package de.skillkiller.documentdbackend.entity.frontend.http.response;
+package de.skillkiller.documentdbackend.entity.http.frontend.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import de.skillkiller.documentdbackend.entity.Document;
